Initial Commit
Initial state of the workbench with as test function written in rust.
This state shows that it is possible to build rust code with setup.py
and to have a FreeCAD workbench together. This is a fundamental design
decision of this worbench and it is needed to work in order for the
workbench to function.