Added Fix Base constraint function.
This constraint function will be used when evaluating the error function
while solving the system of constraints. However, the solver still needs
to be implemented and other constraint functions need to be
implemented. The features added by this commit are the basis for other
constraint functions. Some preliminary experimentation shows that this
fix base constraint function is about 20 times faster than the python
implementation. More performance can be gained by changing from
hyperduals to duals and to create the rotation quaternions outside of
the constraint function.