Added Lock Constraint
Lock constraint is another type of implicit constraint which doesn't
require the use of the solver directly. It simply eliminates the degrees
of freedom of the locked variables. Other constraints need to know about
the locked variables in order to determine partial derivatives since
locked variables are treated as constant values. Note that this
implementation of lock constraints assumes that there will not be any
chain of equality constraints with some chained variable locked. The
assumption is that the solver front end (the part that simplifies and
tailors the system of constraints) will eliminate chained equality
constraints into simpler equality constraints. And if one of them is
locked, then all the equality constraints are also locked.