Moved quaternion building outside of constraint functions.
With this change, quaternions are built outside constraint functions
which removes a great amount of unecessary calculations. Building
quaternions is expensive and they only need to be built a couple of
times; however, before they were built at each iteration of the fix
constraint function. These changes bring up to 30% performance increase
on the small benchmarks I have. However, quaternions may be duplicated
since they are stored in a vector without a way of checking whether
there is already a quaternion representing the rotation of an object.
Further changes to the way the solver treats objects will be needed in
order to avoid this problem of duplicated quaternions.