Added SystemObjects to constraint system solver.
SystemObjects represent objects in 3D space and they contain most of the
data and functions related to the manipulation of the objects they
represent. They contain the variables that represent the position and
rotation of the object (variables were moved from a vector to data
stored in the object). They also contain information about getting a
vector representing the object position and a quaternion representing
the object rotation. These parameters move the logic about making the
vectors and quaternions to outside the constraint functions so that
vector and quaternion creation is not unnecessary done through the
solving process.